+++ src/Pages/Dashboard/Employess/Components/EmployeeStatistics.js
@@ -1,4 +1,4 @@
-import { BiBarChartAlt } from 'react-icons/bi';
+import { BiBarChart } from 'react-icons/bi';
 
 const EmployeeStatistics = ({ employees }) => {
   return (
@@ -8,21 +8,21 @@ const EmployeeStatistics = ({ employees }) => {
           <p className='text-gray-500'>Total Employee</p>
           <p className='text-3xl font-medium text-gray-700'>{employees.length}</p>
         </div>
-        <BiBarChartAlt className='text-5xl text-gray-700' />
+        <BiBarChart className='text-5xl text-gray-700' />
       </div>
       <div className='bg-white py-6 px-10 rounded-lg flex items-center justify-between'>
         <div>
           <p className='text-gray-500'>Male</p>
           <p className='text-3xl font-medium text-gray-700'>145</p>
         </div>
-        <BiBarChartAlt className='text-5xl text-gray-700' />
+        <BiBarChart className='text-5xl text-gray-700' />
       </div>
       <div className='bg-white py-6 px-10 rounded-lg flex items-center justify-between'>
         <div>
           <p className='text-gray-500'>Female</p>
           <p className='text-3xl font-medium text-gray-700'>80</p>
         </div>
-        <BiBarChartAlt className='text-5xl text-gray-700' />
+        <BiBarChart className='text-5xl text-gray-700' />
       </div>
     </div>
   );
